udpate unit tests for emitter
