+++ pkg/emitter/nats_emitter_test.go
@@ -20,12 +20,10 @@ import (
 	"errors"
 	"fmt"
 	"testing"
-	"time"
 
 	jsoniter "github.com/json-iterator/go"
 
 	uuid "github.com/gofrs/uuid"
-	"github.com/guacsec/guac/internal/testing/dochelper"
 	nats_test "github.com/guacsec/guac/internal/testing/nats"
 	"github.com/guacsec/guac/pkg/handler/processor"
 	"github.com/guacsec/guac/pkg/logging"
@@ -78,125 +76,6 @@ var (
 		},
 	}
 )
-
-func TestNatsEmitter_PublishOnEmit(t *testing.T) {
-	expectedDocTree := dochelper.DocNode(&ite6SLSADoc)
-
-	natsTest := nats_test.NewNatsTestServer()
-	url, err := natsTest.EnableJetStreamForTest()
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer natsTest.Shutdown()
-
-	ctx := context.Background()
-	jetStream := NewJetStream(url, "", "")
-	if err := jetStream.JetStreamInit(ctx); err != nil {
-		t.Fatalf("unexpected error initializing jetstream: %v", err)
-	}
-	err = jetStream.RecreateStream(ctx)
-	if err != nil {
-		t.Fatalf("unexpected error recreating jetstream: %v", err)
-	}
-	defer jetStream.Close()
-
-	pubsub := NewEmitterPubSub(ctx, "mem://")
-	ctx = WithEmitter(ctx, pubsub)
-
-	err = testPublish(ctx, &ite6SLSADoc)
-	if err != nil {
-		t.Fatalf("unexpected error on emit: %v", err)
-	}
-
-	var cancel context.CancelFunc
-
-	ctx, cancel = context.WithTimeout(ctx, time.Second)
-	defer cancel()
-
-	transportFunc := func(d processor.DocumentTree) error {
-		if !dochelper.DocTreeEqual(d, expectedDocTree) {
-			t.Errorf("doc tree did not match up, got\n%s, \nexpected\n%s", dochelper.StringTree(d), dochelper.StringTree(expectedDocTree))
-		}
-		return nil
-	}
-
-	err = testSubscribe(ctx, transportFunc)
-	if err != nil {
-		if err != nil && !errors.Is(err, context.DeadlineExceeded) {
-			t.Errorf("nats emitter Subscribe test errored = %v", err)
-		}
-	}
-}
-
-func TestNatsEmitter_PublishOnEmit_DeDuplication(t *testing.T) {
-	expectedDocTree := dochelper.DocNode(&ite6SLSADoc)
-
-	natsTest := nats_test.NewNatsTestServer()
-	url, err := natsTest.EnableJetStreamForTest()
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer natsTest.Shutdown()
-
-	ctx := context.Background()
-	jetStream := NewJetStream(url, "", "")
-	if err := jetStream.JetStreamInit(ctx); err != nil {
-		t.Fatalf("unexpected error initializing jetstream: %v", err)
-	}
-	err = jetStream.RecreateStream(ctx)
-	if err != nil {
-		t.Fatalf("unexpected error recreating jetstream: %v", err)
-	}
-	defer jetStream.Close()
-
-	pubsub := NewEmitterPubSub(ctx, "mem://")
-	ctx = WithEmitter(ctx, pubsub)
-
-	// publish document once
-	err = testPublish(ctx, &ite6SLSADoc)
-	if err != nil {
-		t.Fatalf("unexpected error on emit: %v", err)
-	}
-
-	// publish same document again to check that data deduplication works
-	err = testPublish(ctx, &ite6SLSADoc)
-	if err != nil {
-		t.Fatalf("unexpected error on emit: %v", err)
-	}
-
-	// publish third time the same document to check that data deduplication works
-	err = testPublish(ctx, &ite6SLSADoc)
-	if err != nil {
-		t.Fatalf("unexpected error on emit: %v", err)
-	}
-
-	var cancel context.CancelFunc
-	ctx, cancel = context.WithTimeout(ctx, 2*time.Second)
-	defer cancel()
-
-	listDocument := []processor.DocumentTree{}
-
-	transportFunc := func(d processor.DocumentTree) error {
-		listDocument = append(listDocument, d)
-		return nil
-	}
-
-	err = testSubscribe(ctx, transportFunc)
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			if len(listDocument) != 1 {
-				t.Error("expected only 1 document fetched")
-			}
-			for _, d := range listDocument {
-				if !dochelper.DocTreeEqual(d, expectedDocTree) {
-					t.Errorf("doc tree did not match up, got\n%s, \nexpected\n%s", dochelper.StringTree(d), dochelper.StringTree(expectedDocTree))
-				}
-			}
-		} else {
-			t.Errorf("nats emitter Subscribe test errored = %v", err)
-		}
-	}
-}
 
 func TestNatsEmitter_RecreateStream(t *testing.T) {
 	natsTest := nats_test.NewNatsTestServer()
